requirements.txt
python-dotenv
langchain
langchain-core
langchain-community
langchain-huggingface
huggingface_hub
tavily-python
streamlit